myapp/requirements.txt
anyio==3.6.2 ; python_version >= '3.8'
ariadne==0.17.1
click==8.1.3 ; python_version >= '3.7'
exceptiongroup==1.2.0 ; python_version < '3.11'
gunicorn
flask==2.2.2
flask-cors==3.0.10
flask-sqlalchemy==3.0.2
flask_alembic
flask-migrate==4.0.0
graphql-core==3.2.3 ; python_version >= '3.6' and python_version < '4'
greenlet==2.0.2 ; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
idna==3.4 ; python_version >= '3.5'
importlib-metadata==6.0.0; python_version < '3.10'
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
joblib==1.2.0 ; python_version >= '3.7'
markupsafe==2.1.2 ; python_version >= '3.7'
numpy
pandas==1.5.3
passlib==1.7.4
psycopg2-binary==2.9.5
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2022.7.1
scikit-surprise
scipy==1.10.0; python_version >= '3.9'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0 ; python_version >= '3.7'
sqlalchemy==1.4.46
starlette==0.23.1; python_version >= '3.8'
surprise==0.1
typing-extensions==4.5 ; python_version >= '3.8'
werkzeug==2.2.2 ; python_version >= '3.8'
zipp==3.12.0 ; python_version >= '3.8'
pytube==15.0
openai-whisper
openai
tiktoken
chromadb-client
ffmpeg-python
flask_executor
sentence_transformers